# feat(outputs.sql): Allow sending batches of metrics in transactions