--- a/plugins/outputs/sql/sample.conf
+++ b/plugins/outputs/sql/sample.conf
@@ -41,6 +41,9 @@
   ## Initialization SQL
   # init_sql = ""
 
+  ## Send metrics with the same columns and the same table as batches using prepared statements
+  # batch_transactions = false
+
   ## Maximum amount of time a connection may be idle. "0s" means connections are
   ## never closed due to idle time.
   # connection_max_idle_time = "0s"