## Summary
The current approach to writing, issues an INSERT query for every metric in the buffer, which while working, leaves room for improvement.

This commit modifies the Write behavior as follows:

- Sort the metrics fields/tags (to ensure greater cache-hit ratio)
- Add Prepared statements option

This option caches multiple metrics, so they are issued as a single query.

The results depends on the metrics buffer, the metric cardinality, the output server configuration, etc.

That being said, in micro-benchmarks and production deployment, we've seen 10-100x improvement in write performance.

Due to a priority shift, the previous PR (#16725) has been closed, this one continues with the suggestion to use a hash in order to cache INSERT query template.

In this PR, we're using XXHash (indirect dependency) as it promises better performance across wide variety of hardware.

INSERT query generation is used by default for both Write paths (Current and Prepared), in order to avoid maintenance burden and simplify the Write logic. It should not affect existing users and provides a minor performance boost for them as well.

### Benchmark:

Batches of 1000, 5000, 10000, 50000 and 100000 metrics.

```
                                   │       Write       │               Prepared               │
                                   │      sec/op       │    sec/op     vs base                │
SqliteThroughput/metrics_1000-4           96.629m ± 1%   6.301m ± 11%  -93.48% (p=0.000 n=10)
SqliteThroughput/metrics_5000-4           479.83m ± 1%   29.68m ±  3%  -93.82% (p=0.000 n=10)
SqliteThroughput/metrics_10000-4          961.55m ± 1%   58.56m ±  1%  -93.91% (p=0.000 n=10)
SqliteThroughput/metrics_50000-4          4831.4m ± 0%   294.8m ±  2%  -93.90% (p=0.000 n=10)
SqliteThroughput/metrics_100000-4         9672.9m ± 0%   586.4m ±  1%  -93.94% (p=0.000 n=10)
geomean                                     1.158        71.69m        -93.81%
```

Note:

There are many flaws with this benchmark: SQLite is not representative of every driver, system-specific conditions, metrics had very low cardinality (fewer INSERT queries).

But while a bad benchmark, it's not completely useless and it shows the stark difference in extreme cases.

## Checklist
- [X] No AI generated code was used in this PR