@srebhan 2025-06-30
@hautecodure what is the benefit of `xxhash` over `fnv`? If there is no significant speed gain I would keep with the **built-in** library rather than adding another 3rd-party library that has the risk of being abandoned in the future...