@srebhan 2025-06-30
For the `slices.SortFunc`: I of course meant `slices.SortedFunc`... That's the price you pay if you make suggestions using the GitHub UI without any compiler looking at it... :-)