@hautecodure 2025-07-01
Thanks for the clarification.

Pushed a new version, using the SortedFunc and iterators, not sure if it's clean enough though. Kinda dislike helper methods.

As far as XXHash goes, i first attempted to use FNV, but ran these Go [benchmarks](https://github.com/OneOfOne/xxhash) and it was faster with the sample (pulled from prod) queries i ran through. Same for the synthetic (one metric per query) queries used in some of the tests.

It's important to mention that:
- Queries should see a benefit, as just "INSERT INTO" is longer than 10bytes and depending on the plugin (thus the number of fields), a query could be over a 100 bytes
- XXHash should provide better collision resistance ([according](https://github.com/Cyan4973/xxHash) to the authors)

That said, if XXhash turns out to be an issue, swapping it for FNV would be a 2-line PR (delete the import and rename the hash inside processMetric) as the API is the same.

The only issue i see is the linter complaining about the import placement (it wants to put it below Clickhouse, thus in the middle of the db driver block) and '//nolint' seemingly does not work for gci