@srebhan 2025-07-01
@hautecodure introducing new libraries always is a hassle as they tend to end in an unmaintained state sooner or later. Looking at https://github.com/smallnest/hash-bench and having tested it with `xxhash/v2` it seems like the [standard maphash](https://pkg.go.dev/hash/maphash) is comparable to xxhash and has the advantage that it is a built-in library. What do you think?