@hautecodure 2025-07-01
Thanks, that looks even better, not sure about the collision resistance, but I'm not an expert on one-way hashing functions to have an opinion on it anyway.

Pushed the new version up, let me know if there's anything else that we could improve.